added testAddExistingUser

+++ src/main/java/edu/wofford/wocoin/Database.java
@@ -1,5 +1,7 @@
 package edu.wofford.wocoin;
 
+import org.junit.Test;
+
 import java.io.*;
 import java.sql.*;
 
@@ -65,15 +67,30 @@ public class Database {
 
                  if (rs.next()) {
                      if(rs.getString(1).equals(id)){
+                         try {
+                             conn.close();
+                         } catch (SQLException e) {
+                             e.printStackTrace();
+                         }
                          return true;
                      }
                      else{
+                         try {
+                             conn.close();
+                         } catch (SQLException e) {
+                             e.printStackTrace();
+                         }
                          return false;
                      }
                  } else {
+                     try {
+                         conn.close();
+                     } catch (SQLException e) {
+                         e.printStackTrace();
+                     }
                     return false;
-
                  }
+
         }
         catch(SQLException e){
             e.printStackTrace();
@@ -106,6 +123,12 @@ public class Database {
                 stmt.setString(3, hash);
                 stmt.executeUpdate();
 
+                try {
+                    conn.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+
                 return true;
             }
             catch(SQLException e){
@@ -120,22 +143,30 @@ public class Database {
     }
 
     public boolean removeUser(String id){
-        try (Connection conn = DriverManager.getConnection(url);
-             PreparedStatement stmt = conn.prepareStatement("DELETE FROM users WHERE id = ?;")){
-            stmt.setString(1,id);
-            stmt.executeUpdate();
-            return true;
-        } catch(SQLException e) {
-            e.printStackTrace();
-            return false; }
-//        }finally {
-//
-//            if (conn != null) {
-//                try {
-//                    conn.close();
-//                } catch (SQLException e) {
-//                }
-//            }
+        if(userExists(id)){
+            try (Connection conn = DriverManager.getConnection(url);
+                 PreparedStatement stmt = conn.prepareStatement("DELETE FROM users WHERE id = ?;")){
+                System.out.println("made connection");
+                stmt.setString(1,id);
+                System.out.println("set id");
+                stmt.executeUpdate();
+                System.out.println("updated");
+
+                try {
+                    conn.close();
+                } catch (SQLException e) {
+                    e.printStackTrace();
+                }
+
+                return true;
+            } catch(SQLException e) {
+                e.printStackTrace();
+                return false; }
+        } else {
+            System.out.println("user doesn't exist");
+            return false;
+        }
+
     }
 
     /**
